name wrapping, tag padding, image-contain And moved markdown.ts
