+++ src/app/directory/[slug]/opengraph-image.tsx
@@ -9,7 +9,7 @@ export const size = { width: 1200, height: 630 };
 export const contentType = 'image/png';
 
 export default async function Image({ params }: { params: { slug: string } }) {
-  const slug = params.slug;
+  const slug = (await params).slug;
 
   const clubData = await db.query.club.findFirst({
     where: (club) => eq(club.slug, slug),
@@ -49,202 +49,209 @@ export default async function Image({ params }: { params: { slug: string } }) {
 
   if (!clubData) {
     return new ImageResponse(
-      (
-        <div
-          style={{
-            position: 'relative',
-            width: '100%',
-            height: '100%',
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-            justifyContent: 'center',
-            color: 'white',
-          }}
-        >
-          {background}
-          <h1>Club Not Found</h1>
-        </div>
-      ),
-      { ...size },
-    );
-  }
-
-  return new ImageResponse(
-    (
       <div
         style={{
           position: 'relative',
           width: '100%',
           height: '100%',
           display: 'flex',
-          flexDirection: 'row',
+          flexDirection: 'column',
           alignItems: 'center',
-          justifyContent: 'space-between',
+          justifyContent: 'center',
           color: 'white',
         }}
       >
         {background}
+        <h1>Club Not Found</h1>
+      </div>,
+      { ...size },
+    );
+  }
 
-        {/* Left Side */}
-        <div
-          style={{ display: 'flex', width: '45%', justifyContent: 'center' }}
-        >
-          {clubData.profileImage && (
-            <div
-              style={{
-                display: 'flex',
-                position: 'relative',
-                alignItems: 'center',
-                justifyContent: 'center',
-                width: 350,
-                height: 350,
-                borderRadius: '20px',
-                backgroundColor: 'white',
-                border: '0px solid white',
-                boxShadow: '0 10px 20px rgba(0,0,0,0.3)',
-                overflow: 'hidden',
-              }}
-            >
-              <img
-                src={clubData.profileImage}
-                alt={clubData.name + ' logo'}
-                style={{
-                  width: '100%',
-                  height: '100%',
-                  objectFit: 'cover',
-                }}
-              />
-            </div>
-          )}
-        </div>
+  return new ImageResponse(
+    <div
+      style={{
+        position: 'relative',
+        width: '100%',
+        height: '100%',
+        display: 'flex',
+        flexDirection: 'row',
+        alignItems: 'center',
+        justifyContent: 'space-between',
+        color: 'white',
+      }}
+    >
+      {background}
 
-        {/* Right Side */}
-        <div
-          style={{
-            display: 'flex',
-            flexDirection: 'column',
-            width: '95%',
-            paddingRight: '40px',
-          }}
-        >
-          <h1
-            style={{
-              fontFamily: 'Bai Jamjuree',
-              fontSize: '60px',
-              fontWeight: 'bold',
-              margin: '0 0 20px 0',
-              lineHeight: 1.1,
-              textShadow: '0 2px 10px rgba(0,0,0,0.2)',
-            }}
-          >
-            {clubData.name}
-          </h1>
-
-          {/* Details Container */}
+      {/* Left Side */}
+      <div style={{ display: 'flex', width: '45%', justifyContent: 'center' }}>
+        {clubData.profileImage && (
           <div
             style={{
               display: 'flex',
-              flexDirection: 'row',
-              justifyContent: 'flex-start',
+              position: 'relative',
               alignItems: 'center',
-              width: '70%',
-              fontFamily: 'Inter',
-              fontSize: '25px',
-              margin: '0 0 20px 0',
-              gap: '12px',
+              justifyContent: 'center',
+              width: 350,
+              height: 350,
+              borderRadius: '20px',
+              border: '0px solid white',
+              boxShadow: '0 10px 20px rgba(0,0,0,0.3)',
+              overflow: 'hidden',
             }}
           >
-            {/* Number of members */}
-            {clubData.userMetadataToClubs.length > 1 && (
-              <div
-                style={{
-                  display: 'flex',
-                  flexDirection: 'row',
-                  gap: '12px',
-                }}
-              >
-                <div
-                  style={{
-                    display: 'flex',
-                    position: 'relative',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    width: 35,
-                    height: 35,
-                  }}
-                >
-                  <img
-                    // @ts-expect-error ArrayBuffers are allowed
-                    src={people_alt_icon_buffer}
-                    alt="people icon"
-                    style={{
-                      width: '100%',
-                      height: '100%',
-                      objectFit: 'contain',
-                    }}
-                  />
-                </div>
-                <div style={{ display: 'flex' }}>
-                  {clubData.userMetadataToClubs.length} members
-                </div>
-                {/* Divider */}
-                <div
-                  style={{
-                    width: '2px',
-                    height: '35px',
-                    backgroundColor: 'white',
-                    margin: '0 10px',
-                  }}
-                />
-              </div>
-            )}
+            <img
+              src={clubData.profileImage}
+              alt={clubData.name + ' logo'}
+              style={{
+                width: '100%',
+                objectFit: 'contain',
+              }}
+            />
+          </div>
+        )}
+      </div>
 
-            {/* Nebula Logo */}
+      {/* Right Side */}
+      <div
+        style={{
+          display: 'flex',
+          flexDirection: 'column',
+          width: '95%',
+          paddingRight: '40px',
+        }}
+      >
+        <h1
+          style={{
+            fontFamily: 'Bai Jamjuree',
+            fontSize: '60px',
+            fontWeight: 'bold',
+            margin: '0 0 20px 0',
+            lineHeight: 1.1,
+            textShadow: '0 2px 10px rgba(0,0,0,0.2)',
+            maxWidth: '55%',
+            overflowWrap: 'break-word',
+          }}
+        >
+          {clubData.name}
+        </h1>
+
+        {/* Details Container */}
+        <div
+          style={{
+            display: 'flex',
+            flexDirection: 'row',
+            justifyContent: 'flex-start',
+            alignItems: 'center',
+            width: '70%',
+            fontFamily: 'Inter',
+            fontSize: '25px',
+            margin: '0 0 20px 0',
+            gap: '12px',
+          }}
+        >
+          {/* Number of members */}
+          {clubData.userMetadataToClubs.length > 1 && (
             <div
               style={{
                 display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-                width: 40,
-                height: 40,
+                flexDirection: 'row',
+                gap: '12px',
               }}
             >
-              <NebulaLogo
+              <div
                 style={{
-                  width: '100%',
-                  height: '100%',
-                  objectFit: 'contain',
+                  display: 'flex',
+                  position: 'relative',
+                  alignItems: 'center',
+                  justifyContent: 'center',
+                  width: 35,
+                  height: 35,
+                }}
+              >
+                <img
+                  // @ts-expect-error ArrayBuffers are allowed
+                  src={people_alt_icon_buffer}
+                  alt="people icon"
+                  style={{
+                    width: '100%',
+                    height: '100%',
+                    objectFit: 'contain',
+                  }}
+                />
+              </div>
+              <div
+                style={{
+                  display: 'flex',
+                  textShadow: '0 2px 10px rgba(0,0,0,0.2)',
+                }}
+              >
+                {clubData.userMetadataToClubs.length} members
+              </div>
+              {/* Divider */}
+              <div
+                style={{
+                  width: '2px',
+                  height: '25px',
+                  backgroundColor: '#d4d4d4',
+                  margin: '0 10px 0 9px',
+                  alignSelf: 'center',
                 }}
               />
             </div>
-            <div style={{ display: 'flex' }}>Check it out on UTD CLUBS</div>
-          </div>
-          {clubData.tags && (
-            <div style={{ display: 'flex', flexWrap: 'wrap', gap: '10px' }}>
-              {clubData.tags.map((tag, index) => (
-                <div
-                  key={index}
-                  style={{
-                    backgroundColor: '#d3caff',
-                    color: '#573dff',
-                    padding: '12px 20px',
-                    borderRadius: '50px',
-                    fontSize: '20px',
-                    fontFamily: 'Inter',
-                    fontWeight: 600,
-                    display: 'flex',
-                    alignItems: 'center',
-                  }}
-                >
-                  {tag}
-                </div>
-              ))}
-            </div>
           )}
+
+          {/* Nebula Logo */}
+          <div
+            style={{
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'center',
+              width: 40,
+              height: 40,
+            }}
+          >
+            <NebulaLogo
+              style={{
+                width: '100%',
+                height: '100%',
+                objectFit: 'contain',
+              }}
+            />
+          </div>
+          <div
+            style={{
+              display: 'flex',
+              textShadow: '0 2px 10px rgba(0,0,0,0.2)',
+            }}
+          >
+            Check out on UTD CLUBS
+          </div>
         </div>
+        {clubData.tags && (
+          <div style={{ display: 'flex', flexWrap: 'wrap', gap: '10px' }}>
+            {clubData.tags.map((tag, index) => (
+              <div
+                key={index}
+                style={{
+                  backgroundColor: '#d3caff',
+                  color: '#573dff',
+                  padding: '12px 20px',
+                  borderRadius: '50px',
+                  fontSize: '20px',
+                  fontFamily: 'Inter',
+                  fontWeight: 600,
+                  display: 'flex',
+                  alignItems: 'center',
+                }}
+              >
+                {tag}
+              </div>
+            ))}
+          </div>
+        )}
       </div>
-    ),
+    </div>,
     {
       ...size,
       fonts: [
