+++ src/components/club/ClubCard.tsx
@@ -4,8 +4,8 @@ import { Skeleton } from '@mui/material';
 import Image from 'next/image';
 import Link from 'next/link';
 import { BaseCard } from '@src/components/common/BaseCard';
-import { convertMarkdownToPlaintext } from '@src/modules/markdown';
 import type { SelectClub as Club } from '@src/server/db/models';
+import { convertMarkdownToPlaintext } from '@src/utils/markdown';
 import JoinButton, { JoinButtonSkeleton } from './JoinButton';
 
 type Props = { club: Club; priority?: boolean; manageView?: boolean };
